pyproject.toml
[project]
name = "mmm"
version = "0.1.0"
description = "Masked Diffusion Models for Molecule Generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "networkx>=3.0",
    "rdkit>=2023.3.1",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "einops>=0.6.0",
    "fcd-torch>=1.0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
